ml-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiapp-ml-service"
version = "0.1.0"
description = "ML service for face detection, OCR and speech recognition."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "pillow==11.3.0",
    "opencv-python==4.12.0.88",
    "pytesseract==0.3.13",
    "openai-whisper==20250625",
    "python-multipart==0.0.20",
    "python-dotenv==1.1.1",
    "face-recognition==1.3.0",
]

[project.optional-dependencies]
dev = [
    "httpx==0.28.1",
    "pytest==8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["app"]
